% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcmanalysis-package.R
\docType{package}
\name{kcmanalysis-package}
\alias{kcmanalysis}
\alias{kcmanalysis-package}
\title{kcmanalysis: Common data operations used in research and analysis}
\description{
Use these functions to help in data wrangling and analysis.
}
\author{
\strong{Maintainer}: Celina Garcia \email{n-cegarcia@kingcounty.gov} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
