% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_generic.R
\name{rename_generic}
\alias{rename_generic}
\title{Rename Values in Column}
\usage{
rename_generic(
  data,
  column,
  pattern_replacements,
  na_action = "keep",
  verbose = FALSE
)
}
\arguments{
\item{data}{A dataframe to process.}

\item{column}{The name of the column to transform.}

\item{pattern_replacements}{A named vector with key:value pairs ex. userlanguage = language, ....}

\item{na_action}{A character string specifying the action for NA values:
"keep" (default), "remove", or a specific value for replacement.}
}
\description{
Rename Values in Column
}
