% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_wgt_collapse.R
\name{survey_collapse}
\alias{survey_collapse}
\title{Collapse Survey Weights and Calculate Proportions}
\usage{
survey_collapse(
  data,
  depvar,
  response,
  wgt,
  grouping_vars = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{A dataframe containing survey data.}

\item{depvar}{Name of the dependent variable column.}

\item{response}{Name of the response variable column.}

\item{wgt}{Name of the weight variable column.}

\item{grouping_vars}{Vector of strings of columns used for additional grouping, NULL if no additional grouping is needed.}

\item{verbose}{Boolean indicating if detailed output should be printed.}
}
\description{
Collapse Survey Weights and Calculate Proportions
}
