% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_reshape.R
\name{survey_reshape}
\alias{survey_reshape}
\title{Reshape Survey Data}
\usage{
survey_reshape(data, pivot_cols)
}
\arguments{
\item{data}{A dataframe containing survey data.}

\item{pivot_cols}{Vector of string column names to pivot into longer format.}
}
\description{
This function reshapes a survey dataframe from wide to long format.
}
