% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_dv_flip.R
\name{viz_dv_flip}
\alias{viz_dv_flip}
\title{Generate a horizontal dot-and-whisker plot with customizable appearance}
\usage{
viz_dv_flip(
  data,
  outcome_var = "prop",
  lower_bound = "prop_low",
  vallabel = "cat",
  upper_bound = "prop_upp",
  label_var = "proplabel",
  ymin = 0,
  ymax = 100,
  lang = "en",
  highlight = "",
  main_title = "",
  source_info = "",
  subtitle = "",
  sort = "",
  color_scheme = "#784885",
  label_size = 5
)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{outcome_var}{The title of the outcome variable column in the data frame.}

\item{lower_bound}{The lower bound of the confidence interval.}

\item{vallabel}{The title of the categorical variable column in the data frame.}

\item{upper_bound}{The upper bound of the confidence interval.}

\item{label_var}{The title of the label variable column in the data frame.}

\item{ymin}{The minimum value for the y-axis.}

\item{ymax}{The maximum value for the y-axis.}

\item{lang}{Language setting for text elements (default: "en").}

\item{highlight}{A value to highlight in the plot.}

\item{main_title}{Title of the plot.}

\item{source_info}{Information about the data source.}

\item{subtitle}{Subtitle of the plot.}

\item{sort}{The sorting order of the data (options: "hi-lo", "lo-hi", "alpha", default: "").}

\item{color_scheme}{Color scheme for the plot (default: "#784885").}

\item{label_size}{Size of the labels.}
}
\value{
A ggplot object representing the horizontal dot-and-whisker plot.
}
\description{
This function generates a horizontal dot-and-whisker plot using ggplot2 with customizable appearance settings.
}
