% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_dv_group.R
\name{viz_dv_group}
\alias{viz_dv_group}
\title{Generate a grouped dot-and-whisker plot with customizable appearance}
\usage{
viz_dv_group(
  data,
  depvar = "depvar",
  outcome_var = "prop",
  lower_bound = "prop_low",
  upper_bound = "prop_upp",
  label_var = "proplabel",
  groupingvar = "groupingvar",
  ymin = 0,
  ymax = 100,
  lang = "en",
  main_title = "",
  source_info = "",
  subtitle = "",
  sort = "",
  y_label = "",
  x_label = "",
  color_scheme = c("#D67619", "#264d5e", "#006848", "#4B2884"),
  label_size = 4.25,
  text_position = 0.75
)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{depvar}{The title of the dependent variable column in the data frame.}

\item{outcome_var}{The title of the outcome variable column in the data frame.}

\item{lower_bound}{The lower bound of the confidence interval.}

\item{upper_bound}{The upper bound of the confidence interval.}

\item{label_var}{The title of the label variable column in the data frame.}

\item{groupingvar}{The title of the grouping variable column in the data frame.}

\item{ymin}{The minimum value for the y-axis.}

\item{ymax}{The maximum value for the y-axis.}

\item{lang}{Language setting for text elements (default: "en").}

\item{main_title}{Title of the plot.}

\item{source_info}{Information about the data source.}

\item{subtitle}{Subtitle of the plot.}

\item{sort}{The sorting order of the data (options: "var1", "var2", "var3", "alpha", default: "").}

\item{y_label}{Label for the y-axis.}

\item{x_label}{Label for the x-axis.}

\item{color_scheme}{Color scheme for the plot (default: c("#D67619", "#264d5e", "#006848", "#4B2884")).}

\item{label_size}{Size of the labels.}

\item{text_position}{Position of the text labels.}
}
\value{
A ggplot object representing the grouped dot-and-whisker plot.
}
\description{
This function generates a grouped dot-and-whisker plot using ggplot2 with customizable appearance settings.
}
