% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_fonts.R
\name{viz_fonts}
\alias{viz_fonts}
\title{Add Google fonts for visualization}
\usage{
viz_fonts()
}
\description{
This function adds Google fonts for visualization purposes using sysfonts and showtext.
}
\keyword{internal}
