% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_hist.R
\name{viz_hist}
\alias{viz_hist}
\title{Generate histogram with customizable appearance}
\usage{
viz_hist(
  data,
  outcome_var = "prop",
  label_var = "proplabel",
  cat_var = "cat",
  ymin = 0,
  ymax = 100,
  lang = "en",
  main_title = "",
  subtitle = "",
  source_info = "",
  order = FALSE,
  color_scheme = "#FDB71A"
)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{outcome_var}{The title of the outcome variable column in the data frame.}

\item{label_var}{The title of the label variable column in the data frame.}

\item{cat_var}{The title of the categorical variable column in the data frame.}

\item{ymin}{The minimum value for the y-axis.}

\item{ymax}{The maximum value for the y-axis.}

\item{lang}{Language setting for text elements (default: "en").}

\item{main_title}{Title of the plot.}

\item{subtitle}{Subtitle of the plot.}

\item{source_info}{Information about the data source.}

\item{order}{Logical value indicating whether to order the data by the outcome variable (default: FALSE).}

\item{color_scheme}{Color scheme for the plot (default: "#FDB71A").}
}
\value{
A ggplot object representing the histogram.
}
\description{
This function generates a histogram plot using ggplot2 with customizable appearance settings.
}
